+++ src/lib/st-modal/st-modal.service.ts
@@ -61,7 +61,8 @@ export class StModalService {
       message: string,
       okButton: string,
       cancelButton: string = '',
-      maxWidth: number = 600
+      maxWidth: number = 600,
+      minWidth: number = 400
    ): Observable<StModalResponse> {
 
       const buttons: StModalButton[] = [{
@@ -85,7 +86,8 @@ export class StModalService {
          messageTitle,
          modalTitle,
          buttons,
-         maxWidth
+         maxWidth,
+         minWidth
       });
    }
 
@@ -153,6 +155,7 @@ export class StModalService {
          message: undefined,
          html: undefined,
          maxWidth: undefined,
+         minWidth: undefined,
          empty: false
       };
       const checkedConfig: StModalConfig = Object.assign({}, defaultConfig, config);